requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
accelerate=0.19.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
anyio=4.2.0=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
ca-certificates=2023.12.12=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
clean-fid=0.1.35=pypi_0
click=8.1.7=pypi_0
clip=1.0=pypi_0
comm=0.2.1=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
datasets=2.16.1=pypi_0
decorator=5.1.1=pypi_0
diffusers=0.17.0=pypi_0
dill=0.3.6=pypi_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.108.0=pypi_0
ffmpy=0.3.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.47.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.10.0=pypi_0
ftfy=6.1.3=pypi_0
future=0.18.3=pypi_0
google-auth=2.25.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
gradio=3.36.1=pypi_0
gradio-client=0.2.7=pypi_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.1=pypi_0
idna=3.6=pypi_0
imageio=2.31.1=pypi_0
importlib-metadata=7.0.1=pypi_0
importlib-resources=6.1.1=pypi_0
install=1.3.5=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.1=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lightning-utilities=0.10.0=pypi_0
linkify-it-py=2.0.2=pypi_0
markdown=3.5.1=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.1.1=pypi_0
open-clip-torch=2.23.0=pypi_0
opencv-python=4.8.1.78=pypi_0
openssl=3.0.12=h7f8727e_0
orjson=3.9.10=pypi_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.3=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.7=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=15.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydeprecate=0.3.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyre-extensions=0.0.23=pypi_0
python=3.8.18=h955ad1f_0
python-box=6.0.2=pypi_0
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytorch-lightning=1.4.2=pypi_0
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.32.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rpds-py=0.15.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.1=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.1=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.32.0.post1=pypi_0
sympy=1.12=pypi_0
taming-transformers=0.0.1=dev_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toolz=0.12.0=pypi_0
torch=1.13.1+cu117=pypi_0
torchaudio=0.13.1+cu117=pypi_0
torchmetrics=0.6.0=pypi_0
torchvision=0.14.1+cu117=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.1=pypi_0
transformers=4.28.1=pypi_0
triton=2.1.0=pypi_0
typing-extensions=4.9.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
uc-micro-py=1.0.2=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.25.0=pypi_0
wcwidth=0.2.12=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.5=h5eee18b_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
